Titled borders,instance variables,functions & organization
Added borders around every component that describes that component's
purpose. Added scrollable JLists for the instance variables and
functions created by the user. Used GridBagLayout LayoutManager to
better organize our workspace. For some reason the turtle state text
area is a bit too small. I am still trying to fix that. Otherwise
everything is up to date. Also, changed the name of the CommandPanel
class to ProgrammingPanel and moved the hyperlink to that side as it
makes more sense to have everything related to the programming -
including command references - in that panel.